refactor(includes): minor changes
- load time output format
- code style
