+++ layout/_partial/head/index.ejs
@@ -40,8 +40,6 @@
         return html;
     }) %>
 
-
-
     <!-- Description -->
     <meta name="description" content="<%= page_descr() %>">
     <!-- Keywords -->
